requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.41.0
djangify==1.0.0
Django==5.0.6
django-cloudinary-storage==0.3.0
idna==3.7
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
whitenoise==6.7.0
